+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: gunjkim <gunjkim@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/10 10:20:02 by gunjkim           #+#    #+#             */
-/*   Updated: 2022/11/10 14:21:00 by gunjkim          ###   ########.fr       */
+/*   Updated: 2022/11/16 15:40:37 by gunjkim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	int				index;
+	size_t			index;
 	unsigned char	*temp;
 
 	index = 0;
